看了条件扩散,除了Unet

+++ DiffusionFreeGuidence/DiffusionCondition.py
@@ -1,4 +1,3 @@
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
@@ -24,7 +23,7 @@ class GaussianDiffusionTrainer(nn.Module):
         self.T = T
 
         self.register_buffer(
-            'betas', torch.linspace(beta_1, beta_T, T).double())
+            'betas', torch.linspace(beta_1, beta_T, T).float())
         alphas = 1. - self.betas
         alphas_bar = torch.cumprod(alphas, dim=0)
 
@@ -57,12 +56,13 @@ class GaussianDiffusionSampler(nn.Module):
         ### w > 0 and label > 0 means guidence. Guidence would be stronger if w is bigger.
         self.w = w
 
-        self.register_buffer('betas', torch.linspace(beta_1, beta_T, T).double())
+        self.register_buffer('betas', torch.linspace(beta_1, beta_T, T).float())
         alphas = 1. - self.betas
         alphas_bar = torch.cumprod(alphas, dim=0)
         alphas_bar_prev = F.pad(alphas_bar, [1, 0], value=1)[:T]
         self.register_buffer('coeff1', torch.sqrt(1. / alphas))
         self.register_buffer('coeff2', self.coeff1 * (1. - alphas) / torch.sqrt(1. - alphas_bar))
+        
         self.register_buffer('posterior_var', self.betas * (1. - alphas_bar_prev) / (1. - alphas_bar))
 
     def predict_xt_prev_mean_from_eps(self, x_t, t, eps):
